importer_service/requirements.txt
aiohttp==0.21.5
python-consul==0.6.0
